LIBHYDRA-430. Added Archelon JWT token instructions
Removed the "FCREPO_AUTH_TOKEN" environment variable from
"archelon/archelon.env" because the value cannot be hard-coded into
the configuration because the JWT_SECRET (from which the
FCREPO_AUTH_TOKEN is derived) is not hard-coded.

Added instructions to the README.md file regarding creating the
"FCREPO_AUTH_TOKEN" JWT token.

https://issues.umd.edu/browse/LIBHYDRA-430